ODY-51: Restructured the code
