--- frontend/odyssey-frontend/src/Components/News/News.jsx
@@ -1,290 +1,0 @@
-import React from 'react';
-import { Box, Heading, Text, Grid, Link } from '@chakra-ui/react';
-import Navbar from '../Navbar/Navbar.jsx';
-import Footer from '../Footer/Footer.jsx';
-
-const News = () => {
-
-  return (
-   <div>
-        <Navbar />
-             <Box p="4">
-                  <Heading as="h1" mb="4" textAlign="center" >Latest News Around Kosovo</Heading>
-                  <Grid templateColumns="repeat(auto-fill, minmax(300px, 1fr))" gap="4">
-
-                      <Box borderWidth="1px" borderRadius="lg" overflow="hidden">
-                        <Box p="4">
-                          <Heading as="h2" size="md" mb="2">Tittle</Heading>
-                          <Text fontSize="sm" mb="2">Politics </Text>
-                          <Text fontSize="sm" color="gray.500"> Test</Text>
-                          <Link isExternal color="blue.500">Read More</Link>
-                        </Box>
-                      </Box>
-                      <Box borderWidth="1px" borderRadius="lg" overflow="hidden">
-                        <Box p="4">
-                          <Heading as="h2" size="md" mb="2">Tittle</Heading>
-                          <Text fontSize="sm" mb="2">Sport</Text>
-                          <Text fontSize="sm" color="gray.500"> Test</Text>
-                          <Link isExternal color="blue.500">Read More</Link>
-                        </Box>
-                      </Box>
-                      <Box borderWidth="1px" borderRadius="lg" overflow="hidden">
-                        <Box p="4">
-                          <Heading as="h2" size="md" mb="2">Tittle</Heading>
-                          <Text fontSize="sm" mb="2">Culture</Text>
-                          <Text fontSize="sm" color="gray.500"> Test</Text>
-                          <Link isExternal color="blue.500">Read More</Link>
-                         </Box>
-                      </Box>
-                      <Box borderWidth="1px" borderRadius="lg" overflow="hidden">
-                        <Box p="4">
-                          <Heading as="h2" size="md" mb="2">Tittle</Heading>
-                          <Text fontSize="sm" mb="2">History </Text>
-                          <Text fontSize="sm" color="gray.500"> Test</Text>
-                          <Link isExternal color="blue.500">Read More</Link>
-                        </Box>
-                      </Box>
-                  </Grid>
-              </Box>
-
-
-             <Box p="4">
-                  <Heading as="h1" mb="4" textAlign="center">Latest News Around Turkey</Heading>
-                  <Grid templateColumns="repeat(auto-fill, minmax(300px, 1fr))" gap="4">
-
-                      <Box borderWidth="1px" borderRadius="lg" overflow="hidden">
-                        <Box p="4">
-                          <Heading as="h2" size="md" mb="2">Tittle</Heading>
-                          <Text fontSize="sm" mb="2">Politics </Text>
-                          <Text fontSize="sm" color="gray.500"> Test</Text>
-                          <Link isExternal color="blue.500">Read More</Link>
-                        </Box>
-                      </Box>
-                      <Box borderWidth="1px" borderRadius="lg" overflow="hidden">
-                        <Box p="4">
-                          <Heading as="h2" size="md" mb="2">Tittle</Heading>
-                          <Text fontSize="sm" mb="2">Sport</Text>
-                          <Text fontSize="sm" color="gray.500"> Test</Text>
-                          <Link isExternal color="blue.500">Read More</Link>
-                        </Box>
-                      </Box>
-                      <Box borderWidth="1px" borderRadius="lg" overflow="hidden">
-                        <Box p="4">
-                          <Heading as="h2" size="md" mb="2">Tittle</Heading>
-                          <Text fontSize="sm" mb="2">Culture</Text>
-                          <Text fontSize="sm" color="gray.500"> Test</Text>
-                          <Link isExternal color="blue.500">Read More</Link>
-                        </Box>
-                      </Box>
-                      <Box borderWidth="1px" borderRadius="lg" overflow="hidden">
-                        <Box p="4">
-                          <Heading as="h2" size="md" mb="2">Tittle</Heading>
-                          <Text fontSize="sm" mb="2">History </Text>
-                          <Text fontSize="sm" color="gray.500"> Test</Text>
-                          <Link isExternal color="blue.500">Read More</Link>
-                        </Box>
-                      </Box>
-                  </Grid>
-              </Box>
-
-
-             <Box p="4">
-                  <Heading as="h1" mb="4" textAlign="center">Latest News Around Italy</Heading>
-                  <Grid templateColumns="repeat(auto-fill, minmax(300px, 1fr))" gap="4">
-
-                      <Box borderWidth="1px" borderRadius="lg" overflow="hidden">
-                        <Box p="4">
-                          <Heading as="h2" size="md" mb="2">Tittle</Heading>
-                          <Text fontSize="sm" mb="2">Politics </Text>
-                          <Text fontSize="sm" color="gray.500"> Test</Text>
-                          <Link isExternal color="blue.500">Read More</Link>
-                        </Box>
-                      </Box>
-                      <Box borderWidth="1px" borderRadius="lg" overflow="hidden">
-                        <Box p="4">
-                          <Heading as="h2" size="md" mb="2">Tittle</Heading>
-                          <Text fontSize="sm" mb="2">Sport</Text>
-                          <Text fontSize="sm" color="gray.500"> Test</Text>
-                          <Link isExternal color="blue.500">Read More</Link>
-                        </Box>
-                      </Box>
-                      <Box borderWidth="1px" borderRadius="lg" overflow="hidden">
-                        <Box p="4">
-                          <Heading as="h2" size="md" mb="2">Tittle</Heading>
-                          <Text fontSize="sm" mb="2">Culture</Text>
-                          <Text fontSize="sm" color="gray.500"> Test</Text>
-                          <Link isExternal color="blue.500">Read More</Link>
-                        </Box>
-                      </Box>
-                      <Box borderWidth="1px" borderRadius="lg" overflow="hidden">
-                        <Box p="4">
-                          <Heading as="h2" size="md" mb="2">Tittle</Heading>
-                          <Text fontSize="sm" mb="2">History </Text>
-                          <Text fontSize="sm" color="gray.500"> Test</Text>
-                          <Link isExternal color="blue.500">Read More</Link>
-                        </Box>
-                      </Box>
-                  </Grid>
-              </Box>
-
-
-             <Box p="4">
-                  <Heading as="h1" mb="4" textAlign="center">Latest News Around Spain</Heading>
-                  <Grid templateColumns="repeat(auto-fill, minmax(300px, 1fr))" gap="4">
-
-                      <Box borderWidth="1px" borderRadius="lg" overflow="hidden">
-                        <Box p="4">
-                          <Heading as="h2" size="md" mb="2">Tittle</Heading>
-                          <Text fontSize="sm" mb="2">Politics </Text>
-                          <Text fontSize="sm" color="gray.500"> Test</Text>
-                          <Link isExternal color="blue.500">Read More</Link>
-                        </Box>
-                      </Box>
-                      <Box borderWidth="1px" borderRadius="lg" overflow="hidden">
-                        <Box p="4">
-                          <Heading as="h2" size="md" mb="2">Tittle</Heading>
-                          <Text fontSize="sm" mb="2">Sport</Text>
-                          <Text fontSize="sm" color="gray.500"> Test</Text>
-                          <Link isExternal color="blue.500">Read More</Link>
-                        </Box>
-                      </Box>
-                      <Box borderWidth="1px" borderRadius="lg" overflow="hidden">
-                        <Box p="4">
-                          <Heading as="h2" size="md" mb="2">Tittle</Heading>
-                          <Text fontSize="sm" mb="2">Culture</Text>
-                          <Text fontSize="sm" color="gray.500"> Test</Text>
-                          <Link isExternal color="blue.500">Read More</Link>
-                        </Box>
-                      </Box>
-                      <Box borderWidth="1px" borderRadius="lg" overflow="hidden">
-                        <Box p="4">
-                          <Heading as="h2" size="md" mb="2">Tittle</Heading>
-                          <Text fontSize="sm" mb="2">History </Text>
-                          <Text fontSize="sm" color="gray.500"> Test</Text>
-                          <Link isExternal color="blue.500">Read More</Link>
-                        </Box>
-                      </Box>
-                  </Grid>
-              </Box>
-
-
-             <Box p="4">
-                  <Heading as="h1" mb="4" textAlign="center">Latest News Around Germany</Heading>
-                  <Grid templateColumns="repeat(auto-fill, minmax(300px, 1fr))" gap="4">
-
-                      <Box borderWidth="1px" borderRadius="lg" overflow="hidden">
-                        <Box p="4">
-                          <Heading as="h2" size="md" mb="2">Tittle</Heading>
-                          <Text fontSize="sm" mb="2">Politics </Text>
-                          <Text fontSize="sm" color="gray.500"> Test</Text>
-                          <Link isExternal color="blue.500">Read More</Link>
-                        </Box>
-                      </Box>
-                      <Box borderWidth="1px" borderRadius="lg" overflow="hidden">
-                        <Box p="4">
-                          <Heading as="h2" size="md" mb="2">Tittle</Heading>
-                          <Text fontSize="sm" mb="2">Sport</Text>
-                          <Text fontSize="sm" color="gray.500"> Test</Text>
-                          <Link isExternal color="blue.500">Read More</Link>
-                        </Box>
-                      </Box>
-                      <Box borderWidth="1px" borderRadius="lg" overflow="hidden">
-                        <Box p="4">
-                          <Heading as="h2" size="md" mb="2">Tittle</Heading>
-                          <Text fontSize="sm" mb="2">Culture</Text>
-                          <Text fontSize="sm" color="gray.500"> Test</Text>
-                          <Link isExternal color="blue.500">Read More</Link>
-                        </Box>
-                      </Box>
-                      <Box borderWidth="1px" borderRadius="lg" overflow="hidden">
-                        <Box p="4">
-                          <Heading as="h2" size="md" mb="2">Tittle</Heading>
-                          <Text fontSize="sm" mb="2">History </Text>
-                          <Text fontSize="sm" color="gray.500"> Test</Text>
-                          <Link isExternal color="blue.500">Read More</Link>
-                        </Box>
-                      </Box>
-                  </Grid>
-              </Box>
-             <Box p="4">
-                  <Heading as="h1" mb="4" textAlign="center">Latest News Around Croatia</Heading>
-                  <Grid templateColumns="repeat(auto-fill, minmax(300px, 1fr))" gap="4">
-
-                      <Box borderWidth="1px" borderRadius="lg" overflow="hidden">
-                        <Box p="4">
-                          <Heading as="h2" size="md" mb="2">Tittle</Heading>
-                          <Text fontSize="sm" mb="2">Politics </Text>
-                          <Text fontSize="sm" color="gray.500"> Test</Text>
-                          <Link isExternal color="blue.500">Read More</Link>
-                        </Box>
-                      </Box>
-                      <Box borderWidth="1px" borderRadius="lg" overflow="hidden">
-                        <Box p="4">
-                          <Heading as="h2" size="md" mb="2">Tittle</Heading>
-                          <Text fontSize="sm" mb="2">Sport</Text>
-                          <Text fontSize="sm" color="gray.500"> Test</Text>
-                          <Link isExternal color="blue.500">Read More</Link>
-                        </Box>
-                      </Box>
-                      <Box borderWidth="1px" borderRadius="lg" overflow="hidden">
-                        <Box p="4">
-                          <Heading as="h2" size="md" mb="2">Tittle</Heading>
-                          <Text fontSize="sm" mb="2">Culture</Text>
-                          <Text fontSize="sm" color="gray.500"> Test</Text>
-                          <Link isExternal color="blue.500">Read More</Link>
-                        </Box>
-                      </Box>
-                      <Box borderWidth="1px" borderRadius="lg" overflow="hidden">
-                        <Box p="4">
-                          <Heading as="h2" size="md" mb="2">Tittle</Heading>
-                          <Text fontSize="sm" mb="2">History </Text>
-                          <Text fontSize="sm" color="gray.500"> Test</Text>
-                          <Link isExternal color="blue.500">Read More</Link>
-                        </Box>
-                      </Box>
-                  </Grid>
-              </Box>
-             <Box p="4">
-                  <Heading as="h1" mb="4" textAlign="center">Latest News Around Swedeen</Heading>
-                  <Grid templateColumns="repeat(auto-fill, minmax(300px, 1fr))" gap="4">
-                      <Box borderWidth="1px" borderRadius="lg" overflow="hidden">
-                        <Box p="4">
-                          <Heading as="h2" size="md" mb="2">Tittle</Heading>
-                          <Text fontSize="sm" mb="2">Politics </Text>
-                          <Text fontSize="sm" color="gray.500"> Test</Text>
-                          <Link isExternal color="blue.500">Read More</Link>
-                        </Box>
-                      </Box>
-                      <Box borderWidth="1px" borderRadius="lg" overflow="hidden">
-                        <Box p="4">
-                          <Heading as="h2" size="md" mb="2">Tittle</Heading>
-                          <Text fontSize="sm" mb="2">Sport</Text>
-                          <Text fontSize="sm" color="gray.500"> Test</Text>
-                          <Link isExternal color="blue.500">Read More</Link>
-                        </Box>
-                      </Box>
-                      <Box borderWidth="1px" borderRadius="lg" overflow="hidden">
-                        <Box p="4">
-                          <Heading as="h2" size="md" mb="2">Tittle</Heading>
-                          <Text fontSize="sm" mb="2">Culture</Text>
-                          <Text fontSize="sm" color="gray.500"> Test</Text>
-                          <Link isExternal color="blue.500">Read More</Link>
-                        </Box>
-                      </Box>
-                      <Box borderWidth="1px" borderRadius="lg" overflow="hidden">
-                        <Box p="4">
-                          <Heading as="h2" size="md" mb="2">Tittle</Heading>
-                          <Text fontSize="sm" mb="2">History </Text>
-                          <Text fontSize="sm" color="gray.500"> Test</Text>
-                          <Link isExternal color="blue.500">Read More</Link>
-                        </Box>
-                      </Box>
-                  </Grid>
-              </Box>
-        <Footer />
-   </div>
-  );
-};
-
-
-export default News;
