+++ frontend/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "odyssey-frontend",
+  "name": "frontend",
   "private": true,
   "version": "0.0.0",
   "type": "module",
@@ -13,7 +13,10 @@
     "@chakra-ui/react": "^2.8.2",
     "@emotion/react": "^11.11.4",
     "@emotion/styled": "^11.11.5",
-    "react": "^18.2.0",
+    "axios": "^1.6.8",
+    "framer-motion": "^11.2.0",
+    "icons": "^1.0.0",
+    "react": "^18.3.1",
     "react-dom": "^18.2.0",
     "react-icons": "^5.2.1",
     "react-router-dom": "^6.23.1"
@@ -26,6 +29,6 @@
     "eslint-plugin-react": "^7.34.1",
     "eslint-plugin-react-hooks": "^4.6.0",
     "eslint-plugin-react-refresh": "^0.4.6",
-    "vite": "^5.2.10"
+    "vite": "^5.2.0"
   }
 }
