+++ frontend/src/assets/activities.jpeg
