+++ frontend/src/assets/news.jpeg
