+++ frontend/src/assets/trips.jpeg
