+++ frontend/src/components/authentication/Authentication.jsx
@@ -1,25 +1,22 @@
-import React, { useState } from 'react';
-import "./LoginSignup.css";
+import { React, useState } from "react";
+import "./Authentication.css";
 
+const Authentication = () => {
+    const [action, setAction] = useState("Login");
 
-// import user_icon from '../Assets/profile.png'
-// import email_icon from '../Assets/email.png'
-// import password_icon from '../Assets/password.png'
-
-const LoginSignup = () => {
-
- const [action,setAction] = useState("Login");
     return(
         <div className='container'>
             <div className="header">
-                <div className="text">{action}</div>
+                <div className="text">{ action }</div>
                 <div className="underline"></div>
             </div>
 
             <div className="inputs">
-            {action==="Login"?<div></div>:<div className="input">
-                                                <input type="text" placeholder="Name" />
-                                           </div>
+            { action==="Login" ? 
+                <div></div> : 
+                <div className="input">
+                    <input type="text" placeholder="Name" />
+                </div>
             }
 
                 <div className="input">
@@ -38,7 +35,7 @@ const LoginSignup = () => {
                 <div className={action==="Sign Up"?"submit gray":"submit"} onClick={()=>{setAction("Login")}}>Login</div>
             </div>
         </div>
-    )
+    );
 }
-
-export default LoginSignup
+   
+export default Authentication;
