+++ frontend/src/components/shared/Navbar.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Box, Flex, Spacer, Link, Heading, Menu, MenuButton, MenuList, MenuItem, Icon } from '@chakra-ui/react';
-import { FaAngleDown } from 'react-icons/fa'; // Import the desired icon
-import logo from "../Assets/logo.png";
+import { FaAngleDown } from 'react-icons/fa';
+import logo from "../../assets/odyssey-logo.png";
 import { Link as RouterLink } from 'react-router-dom';
 
 const Navbar = () => {
@@ -13,9 +13,9 @@ const Navbar = () => {
             <Spacer />
             <Box mr={{ base: '0', md: '400px' }} mt={{ base: '4', md: '0' }} display={{ base: 'none', md: 'block' }}>
                 <RouterLink to="/"><Link mr="8" _hover={{ textDecoration: 'none', opacity:'50%' }}>Home</Link></RouterLink>
-                <RouterLink to="/contactus"><Link  mr="8" _hover={{ textDecoration: 'none', opacity:'50%' }}>Contact Us</Link></RouterLink>
+                <RouterLink to="/contact"><Link  mr="8" _hover={{ textDecoration: 'none', opacity:'50%' }}>Contact Us</Link></RouterLink>
                 <RouterLink ><Link mr="8" _hover={{ textDecoration: 'none', opacity:'50%' }}>Destination</Link></RouterLink>
-                <RouterLink to="/aboutus"><Link mr="8" _hover={{ textDecoration: 'none', opacity:'50%' }}>About Us</Link></RouterLink>
+                <RouterLink to="/about-us"><Link mr="8" _hover={{ textDecoration: 'none', opacity:'50%' }}>About Us</Link></RouterLink>
                 <Menu>
                     <MenuButton as={Link} mr="4" fontSize="lg" fontWeight="bold" _hover={{ textDecoration: 'none', opacity:'50%' }}>
                         <Icon as={FaAngleDown} />
@@ -32,7 +32,7 @@ const Navbar = () => {
                     </MenuList>
                 </Menu>
             </Box>
-            <RouterLink to="/loginsignup"><Link mr={{ base: '0', md: '6' }} mt={{ base: '2', md: '0' }} pr="10" _hover={{ textDecoration: 'none', opacity:'50%' }}>Log In</Link></RouterLink>
+            <RouterLink to="/authenticate"><Link mr={{ base: '0', md: '6' }} mt={{ base: '2', md: '0' }} pr="10" _hover={{ textDecoration: 'none', opacity:'50%' }}>Log In</Link></RouterLink>
         </Flex>
     );
 };
