+++ frontend/src/main.jsx
@@ -5,10 +5,11 @@ import { ChakraProvider } from '@chakra-ui/react'
 import './index.css'
 
 ReactDOM
-.createRoot(document.getElementById('root')).render(
-  <React.StrictMode>
-    <ChakraProvider>
-    <App />
-    </ChakraProvider>
-  </React.StrictMode>,
-)
+  .createRoot(document.getElementById('root'))
+  .render(
+    <React.StrictMode>
+      <ChakraProvider>
+        <App />
+      </ChakraProvider>
+    </React.StrictMode>,
+  )
